Added flyer class and now there's amount of flyers that were printed
