+++ Assets/Localization/OldLocalisation/Tables/uiText Shared Data.asset
@@ -364,6 +364,10 @@ MonoBehaviour:
     m_Key: flyer-finished
     m_Metadata:
       m_Items: []
+  - m_Id: 129710360383488
+    m_Key: flyer-amountPrinted
+    m_Metadata:
+      m_Items: []
   m_Metadata:
     m_Items: []
   m_KeyGenerator:
