+++ Assets/Localization/OldLocalisation/Tables/uiText_ro.asset
@@ -129,6 +129,10 @@ MonoBehaviour:
     m_Localized: "Am terminat toate flutura\u0219ii!"
     m_Metadata:
       m_Items: []
+  - m_Id: 129710360383488
+    m_Localized: "Cantitatea flutura\u0219i tip\u0103ri\u021Bi:"
+    m_Metadata:
+      m_Items: []
   references:
     version: 1
     00000000:
