+++ Assets/Scripts/Flyer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 978e5086cc06f1e46837313fc21b4eaa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
