+++ Assets/Scripts/Flyer/Flyer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2dbd0bdbf72acab4c9fe909dad27906d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
