+++ Assets/Scripts/Flyer/FlyerButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b89e4b6c702079b4fa0bb8f21035af25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
