+++ Assets/Scripts/Flyer/FlyerOption.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FlyerOption : MonoBehaviour
+{
+    public float Points;
+}
