+++ Assets/Scripts/Flyer/Printer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2bc0e8d2940965459f145038293524b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
