--- Assets/Scripts/FlyerMaking.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 01ece94f390b2a44bbb93bf647bc72ac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
