--- Assets/Scripts/FlyerOption.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bbab1191322f5c944baccea471f28958
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
